+++ src/Exceptions/ValidationException.php
@@ -1,0 +1,22 @@
+<?php
+
+/*
+ * This file is part of Caspeco.
+ *
+ (c) Schimpanz Solutions <info@schimpanz.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Schimpanz\Caspeco\Exceptions;
+
+/**
+ * This is the validation exception class.
+ *
+ * @author Vincent Klaiber <vincent@schimpanz.com>
+ */
+class ValidationException extends HttpException
+{
+    //
+}