+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.2.0 (upcoming)
 
 - Better error handling
+- Added validation handler
 
 ## 0.1.0 (released 2015-10-13)